Extracted main menu dialog out into own dialog class. Simplified creation of state accessor for dialog set.
