+++ csharp_dotnetcore/wwwroot/default.htm
@@ -392,7 +392,7 @@
     <header class="header">
         <div class="header-inner-container">
             <div class="header-icon" style="display: inline-block"></div>
-            <div class="header-text" style="display: inline-block">food_bot Bot</div>
+            <div class="header-text" style="display: inline-block">FoodBot Bot</div>
         </div>
     </header>
     <div class="row">
